requirements.txt
fastapi
uvicorn

langchain>=1.2.15
langchain-core>=1.2.26
langchain-community>=0.4.1
langchain-text-splitters>=1.1.1

langchain-qdrant
langchain-groq>=1.1.2
langchain-huggingface>=1.2.1

qdrant-client
sentence-transformers>=2.2.2
pypdf
python-dotenv>=1.2.2

ragas
langsmith
flashrank

streamlit
requests
ipykernel>=7.2.0